[FIX] product_supplier_sale_price: One file per model. Applying multi recordset on compute field. And minor fixes for travis.
