+++ product_supplier_sale_price/models/__init__.py
@@ -2,4 +2,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import res_config_settings
-from . import product
+from . import product_product
+from . import product_supplierinfo
