+++ product_supplier_sale_price/readme/DESCRIPTION.rst
@@ -1,1 +1,6 @@
-Set Sale Price using prices suggested by supplier
+Supplier can suggest a sale price for a product
+You can set this suggested sale price as your sale price.
+
+If a product has multiple supplier suggesting a sale price you can decide how to apply the computation of your own sale price:
+get the cheapest, the more expensive, compute the mean or not to apply the suggested sale price to your sale price.
+(how to compute the sale price is determined by the system parameter: Supplier Sale Price Computation)
